backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
kg-gen>=0.1.0
python-dotenv==1.0.1
werkzeug>=3.0.0
google-generativeai==0.3.2 